refactor: mover funcion de enviar token al servidor al API Service y borrar probar token
